+++ lib/math/math.go
@@ -1,0 +1,11 @@
+package math
+
+func MaxUint64(values ...uint64) (max uint64) {
+	for _, v := range values {
+		if v < max {
+			continue
+		}
+		max = v
+	}
+	return max
+}